requirements.txt
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
cryptography==35.0.0
Flask==2.0.1
gunicorn==20.1.0
google-api-core==2.0.1
google-auth==2.2.0
google-cloud-pubsub==2.8.0
google-cloud-bigquery==2.28.1
google-crc32c==1.3.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.41.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
libcst==0.3.21
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.0
proto-plus==1.19.0
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.7
Werkzeug==2.0.1
